Fixed compile error for NativeNativeSoundProvider.kt due to Kotlin 1.5 outlawing Double -> Short conversion (#236)
https://youtrack.jetbrains.com/issue/KT-30360

+++ korau/src/linuxX64Main/kotlin/com/soywiz/korau/sound/NativeNativeSoundProvider.kt
@@ -336,8 +336,8 @@ private fun applyStereoPanningInline(interleaved: ShortArray, panning: Double = 
     val lratio = (1.0 - rratio) * vvolume
     //println("panning=$panning, lratio=$lratio, rratio=$rratio, vvolume=$vvolume")
     for (n in interleaved.indices step 2) {
-        interleaved[n + 0] = (interleaved[n + 0] * lratio).toShort()
-        interleaved[n + 1] = (interleaved[n + 1] * rratio).toShort()
+        interleaved[n + 0] = (interleaved[n + 0] * lratio).toInt().toShort()
+        interleaved[n + 1] = (interleaved[n + 1] * rratio).toInt().toShort()
     }
 }
